nlp: add combined voice+NLU API
Pass a voice file + metadata (developer key, context, etc.) and
retrieve both the text and the corresponding ThingTalk code.

This reduces the latency of Almond replies, as it reduces the
number of round-trips we need.
